+++ FlatFiles/ByteColumn.cs
@@ -40,9 +40,10 @@ namespace FlatFiles
         /// <summary>
         /// Parses the given value into a byte.
         /// </summary>
+        /// <param name="context">Holds information about the column current being processed.</param>
         /// <param name="value">The value to parse.</param>
         /// <returns>The parsed byte value.</returns>
-        public override object Parse(string value)
+        public override object Parse(IColumnContext context, string value)
         {
             if (Preprocessor != null)
             {
@@ -59,9 +60,10 @@ namespace FlatFiles
         /// <summary>
         /// Formats the given object.
         /// </summary>
+        /// <param name="context">Holds information about the column current being processed.</param>
         /// <param name="value">The object to format.</param>
         /// <returns>The formatted value.</returns>
-        public override string Format(object value)
+        public override string Format(IColumnContext context, object value)
         {
             if (value == null)
             {